+++ integration_tests/models/account_view.sql
@@ -1,3 +1,2 @@
 {{ salesforce_formula_utils.sfdc_formula_view(
-    source_table = 'account',
-    full_statement_version = false ) }}
+    source_table = 'account') }}