+++ integration_tests/models/user_statement_view.sql
@@ -1,4 +1,3 @@
 {{ salesforce_formula_utils.sfdc_formula_view(
-    source_table='user',
-    full_statement_version=true) 
+    source_table='user') 
 }}