+++ integration_tests/models/user_view.sql
@@ -1,3 +1,2 @@
 {{ salesforce_formula_utils.sfdc_formula_view(source_table='user',
-    reserved_table_name='user_table',
-    full_statement_version = false ) }}
+    reserved_table_name='user_table') }}